Feat/galvez

+++ services/nutritionist-service/src/main/java/pe/edu/upc/nutritionist_service/nutritionist/interfaces/rest/transform/NutritionistPatientResourceFromEntityAssembler.java
@@ -7,7 +7,7 @@ public class NutritionistPatientResourceFromEntityAssembler {
 
     public static NutritionistPatientResource toResourceFromEntity(NutritionistPatient np) {
         return new NutritionistPatientResource(
-                np.getId(),
+                np.getId().intValue(),
                 np.getNutritionistId(),
                 np.getPatientUserId(),
                 np.getServiceType(),
